Applied some comments
